don't read uninitialized values when changing resolution
When changing the window resolution, the game state is changed to
'null', calling resol_leave() and free'ing the mode-list. Then its value
is passed to video_mode(), resulting in garbage and wrong window size.
The width and height values are saved to local temporary variables
before changing the state, thus passed values are now correct.

This partially fixes issue #3.